backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
mysql-connector-python>=9.0.0
pydantic>=2.9.2